consumer/requirements.txt
kafka-python
pytesseract
psycopg2-binary
sqlalchemy
python-dotenv
pdf2image
pillow

